# docs/swagger.yaml
definitions:
  dto.SignInRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  dto.SignUpRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - phone
    - username
    type: object
  errorresponse.ErrorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service for routing requests to different API endpoints and gateway.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Api-gateway-SiteZtta
  version: "1.0"
paths:
  /auth/test:
    get:
      responses: {}
      security:
      - BearerAuth: []
      summary: Test auth
  /auth/v1/sign-in/:
    post:
      consumes:
      - application/json
      description: Sign in user in system
      operationId: sign-in
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: id
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
      summary: SignInV1
      tags:
      - auth
  /auth/v1/sign-up/:
    post:
      consumes:
      - application/json
      description: Create a new user
      operationId: sign-up
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: id
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/errorresponse.ErrorResponse'
      summary: SignUpV1
      tags:
      - auth
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"